feature: 对其长连接+转发

+++ internal/repository/post.go
@@ -8,7 +8,7 @@ import (
 type PostRepository interface {
 	GetPost(ctx context.Context, id int64) (*model.Post, error)
 	Create(post *model.Post) error
-	GetPostById(postId int64) (*model.Post, error)
+	GetPostById(postId uint) (*model.Post, error)
 	PostPagination(pageNum int, pageSize int) ([]model.Post, error)
 }
 
@@ -34,7 +34,7 @@ func (r *postRepository) Create(post *model.Post) error {
 	return r.db.Create(post).Error
 }
 
-func (r *postRepository) GetPostById(postId int64) (*model.Post, error) {
+func (r *postRepository) GetPostById(postId uint) (*model.Post, error) {
 	var post model.Post
 	result := r.db.Where("id = ?", postId).First(&post)
 	if result.Error != nil {
